requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.3
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
biopython==1.64
celery==3.1.14
itsdangerous==0.24
kombu==3.0.22
pytz==2014.7
redis==2.10.3
reportlab==3.1.8
wsgiref==0.1.2
